refactor: VFolder ID based APIs
